Bio
Elena Zlokazova is a San Francisco-based illustrator and contemporary artist. 
She started off her career working as a fashion designer, later turning her hand to Illustration. Elena specialises in creating artwork that captures the focus in a way that engages the viewer and contains strong character relationships. 
Her artworks were exhibited in Moscow, London, San Francisco and at the Ultramarine Fashion Illustration exhibition in Miami. She was a finalist for the FIDA (Fashion Illustration Drawing Awards) awards in 2020 and 2021.